requirements.txt
tensorflow
polars
numpy
scikit-learn
pandas
fastapi
uvicorn
faiss-cpu
streamlit
requests
pydantic